requirements.txt
together
python-dotenv
pymongo